# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cambridge AI Tutor API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to returnTo on success.
...
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
...
  /admin/plans:
    get:
      operationId: adminListPlans
      tags: [admin]
      summary: List all plans, including inactive ones (admin only)
      responses:
        "200":
          description: All plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPlan"
        "403":
          description: Admin access required
  /admin/plans/{id}:
    put:
      operationId: adminUpdatePlan
      tags: [admin]
      summary: Update a plan's price, limits, top-up eligibility, or visibility (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPlanUpdateBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPlan"
        "403":
          description: Admin access required
        "404":
          description: Unknown plan id
  /admin/credit-packs:
    get:
      operationId: adminListCreditPacks
      tags: [admin]
      summary: List all credit packs, including inactive ones (admin only)
      responses:
        "200":
          description: All credit packs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCreditPack"
        "403":
          description: Admin access required
    post:
      operationId: adminCreateCreditPack
      tags: [admin]
      summary: Create a credit pack (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreditPackInput"
      responses:
        "201":
          description: Created credit pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCreditPack"
        "403":
          description: Admin access required
  /admin/credit-packs/{id}:
    put:
      operationId: adminUpdateCreditPack
      tags: [admin]
      summary: Update a credit pack (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreditPackInput"
      responses:
        "200":
          description: Updated credit pack
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCreditPack"
        "403":
          description: Admin access required
        "404":
          description: Unknown credit pack id
    delete:
      operationId: adminDeleteCreditPack
      tags: [admin]
      summary: Delete a credit pack (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Admin access required
        "404":
          description: Unknown credit pack id
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isAdmin]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        isAdmin:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    SubscriptionFeature:
      type: string
      description: A quota-metered AI feature.
      enum: [tutorMessage, noteGeneration, quizGeneration, flashcardGeneration, oralPractice]
...
    SubscriptionPlanLimit:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/SubscriptionFeature"
        limit:
          type: integer
          description: Units included per billing period for this feature on this plan.
      required: [feature, limit]
    SubscriptionPlan:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        priceCents:
          type: integer
          description: Monthly price in cents (0 = free).
        allowsTopups:
          type: boolean
        limits:
          type: array
          items:
            $ref: "#/components/schemas/SubscriptionPlanLimit"
      required: [id, name, priceCents, allowsTopups, limits]
    AdminPlan:
      allOf:
        - $ref: "#/components/schemas/SubscriptionPlan"
        - type: object
          properties:
            sortOrder:
              type: integer
            isActive:
              type: boolean
          required: [sortOrder, isActive]
    AdminPlanUpdateBody:
      type: object
      properties:
        name:
          type: string
        priceCents:
          type: integer
          minimum: 0
        allowsTopups:
          type: boolean
        sortOrder:
          type: integer
        isActive:
          type: boolean
        limits:
          type: object
          description: Partial map of feature -> new limit; omitted features keep their current limit.
          additionalProperties:
            type: integer
            minimum: 0
    CreditPack:
      type: object
      properties:
        id:
          type: integer
        feature:
          $ref: "#/components/schemas/SubscriptionFeature"
        label:
          type: string
        amount:
          type: integer
        priceCents:
          type: integer
      required: [id, feature, label, amount, priceCents]
    AdminCreditPack:
      allOf:
        - $ref: "#/components/schemas/CreditPack"
        - type: object
          properties:
            isActive:
              type: boolean
          required: [isActive]
    AdminCreditPackInput:
      type: object
      properties:
        feature:
          $ref: "#/components/schemas/SubscriptionFeature"
        label:
          type: string
        amount:
          type: integer
          minimum: 1
        priceCents:
          type: integer
          minimum: 0
        isActive:
          type: boolean
...